LayoutTests/http/tests/w3c/html/semantics/embedded-content/the-iframe-element/iframe_sandbox_allow_script-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'http://127.0.0.1:8000/w3c/html/semantics/embedded-content/the-iframe-element/resources/sandbox_allow_script.html' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.

PASS iframe_sandbox_allow_scripts 
